.PythonInit {RSPython}	R Documentation

Initialize the Python interpreter.

Description

These functions allow one to initializes the Python engine and creates an interpreter and also to discard it and its contents. The .PythonInit function must be called before invoking a Python function or method. One can start, stop and restart the Python interpreter any number of times.

Usage

.PythonInit(path=NULL, merge=T)
.PythonTerminate()

Details

.PythonInit passes its arguments to .PythonPath to determine the appropriate value for the PYTHONPATH environment variable. It then uses the returned value to set this environment variable and then initializes the python interpreter.

Value

NULL or an error occurs.

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

See Also

.PythonPath

Examples

  .PythonInit()

[Package Contents]