Python Reflectance {RSPython}	R Documentation

Information about Python classes and methods

Description

These functions return information about a python class and its methods. From this we

Usage

pythonMethods(obj, all=F)
getSuperClasses(obj, all=F)
pythonModuleNames(obj, all=F)

Arguments

obj 	 a character vector giving the name of a Python class, given in the form c("className", "moduleName")

all 	 controls how inherited methods and attributes are processed. If TRUE is specified, the inherited methods are also reported. Otherwise, these are ignored and only those methods defined within the class are returned.

Details

This uses the code

Value

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

See Also

Examples

[Package Contents]