getPythonReferenceIDs {RSPython}	R Documentation

Python Foreign Reference Information

Description

This returns the names of the currently active foreign Python reference objects being managed by the Python reference manager.

Usage

getPythonReferenceIDs()

Details

This probes the reference manager in which the exported Python objects are actually stored and gets information about each element.

Value

A character vector giving the identifiers (i.e. names) of the currently active foreign references exported from the Python session.

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

See Also

.Python, .PythonNew

Examples

  before <- getPythonReferenceIDs()
   # do something that creates Python reference objects
  after <-  getPythonReferenceIDs()
   # Now cleanup the new ones.

[Package Contents]