importPythonModule {RSPython}	R Documentation

Loads a Python Module

Description

This loads the Python module into the Python interpreter, making its functions and classes available to the Python, and hence R, session. This is equivalent to the Python command from <name> import *.

Usage

importPythonModule(name, all=T)

Arguments

name

all

Value

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

See Also

pythonModuleTypes

Examples

  importPythonModule("sys")
  .PythonEval("version")

  importPythonModule("sys", all=F)
  .PythonEval("sys.version")

  importPythonModule("urllib", all=F)
  pythonModuleTypes("urllib")

[Package Contents]