isPythonInitialized {RSPython}	R Documentation

Is Python Interpreter Initialized

Description

This function tests whether the Python interpreter is in an initialized state and optionally initializes it if it is not (using default settings).

Usage

isPythonInitialized(init=F)

Arguments

init 	 a logical value indicating whether to perform the default initialization if the Python interpreter is not already initialized.

Details

This checks the internal state of the Python library and queries if the interpreter is already initialized. If not and init is TRUE, the function .PythonInit is called with no arguments.

Value

A logical value indicating what the current status of the Python interpreter is. If init is FALSE, this is the state before the call. Otherwise, the interpreter should have been started, except if an error occurs.

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

Examples

 isPythonInitialized(init)

 if(isPythonInitialized())

[Package Contents]