pythonModuleTypes {RSPython}	R Documentation

Python Module Elements and Types

Description

This function provides reflective information about the contents of a module by returning a list of all the entries in that module and the type of object it (currently) is. This allows one to differentiate the functions or methods and the fields of a module and to understand what the fields contain. Essentially it describes the elements of the module's dictionary (\_\_dict\_\_)

Usage

pythonModuleTypes(name)

Arguments

name 	 the name of the module in question. This is given as a single string or a character vector containg the hierarchy of module names in a top-down or top-first form.

Details

Value

A named character/string vector whose name-value pairs give the name of each entry in the module and its type. The types are

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

See Also

Examples

 pythonModuleTypes("sys")
 pythonModuleTypes("RS")

[Package Contents]