referenceManager {RSPython}	R Documentation

Create instance of the standard Reference Manager

Description

Usage

referenceManager()

Details

If no reference manager is available when it is first needed by a call from Python to R, this function is called to create a referenceManager instance and is then registered at the C level.

Value

A closure that stores R objects and returns references to them. The closure provides a variety of different functions to operate on these R objects:

handler

createReference

addReference

remove

getReference

references

total

Author(s)

Duncan Temple Lang

References

http://www.omegahat.org/RSPython, http://www.python.org

See Also

setPythonHandler getPythonHandler

Examples

[Package Contents]